requirements.txt
tweepy>=4.0.0
textblob>=0.15.3
nltk>=3.6.0
python-dotenv>=0.19.0
owlready2>=0.37 